Extract #initialise_globals helper
Unfortunately, because a global variable’s intialiser expression can
potentially refer to a function (via `ref.func`) or an imported global
(via `global.get`) within the same module, we must set the current
module to the module containing the globals being initialised so that we
can successfully evaluate expressions in the context of that module.

I suspect we’ll need to tackle the idea of “evaluate this expression in
the context of this module” in order to successfully evaluate the bodies
of imported WebAssembly (vs native) functions, but we haven’t had to
deal with that yet, so for now we just relying on mutating
`current_module` to keep track of that context.