Set current module to `nil` during construction of new module
This will prevent confusion when module instantiation code accidentally
runs too early and tries to evaluate an expression in the context of a
module which does not yet exist. Before this change it was easy to be
confused when this happened because `current_module` would still be
pointing at the previously-constructed module, which probably has
nothing to do with the current one. At least when it’s `nil` the problem
will be more obvious.

Ultimately we need to do a better job of separating out the two stages
of a) creating a module and b) instantiating it (e.g. setting all the
globals). At the moment they’re both mixed together, but hopefully this
change will make it easier to spot the places where they must be teased
apart.