Separate the stages of building a memory and initialising it with data
As with globals, there’s a chicken-and-egg problem here: memory is part
of a module so we can’t create a module without one, but the process of
initialising memory involves evaluating data offset expressions, which
needs to happen in the context of the current module (e.g. so that
global variables can be read). And as with globals, the solution is to
build an empty memory as part of the module and then fill it in after
the Module instance has been created.