Use `context.tables` when parsing table index in `call_indirect`
There’s another place in the parser (in #parse_export) where we would
like to call #parse_index with `context.tables`, but because the tables
index space is not yet populated, doing so there will break a
currently-passing test. The change in this commit only affects
select.wast, which is already pending.