Support `externref` in table type
I don’t really understand what `externref` means — is it a completely
opaque reference to some object originating from the host environment,
or are there some operations we’re able to perform on it? Anyway,
doesn’t matter for now; we just have to be able to parse it in
elem.wast.