Implement the `ref.func` instruction
At the moment we just push the function _index_ onto the operand stack,
since that’s what the implementation of `call_indirect` expects to find
in the table, but I don’t see any reason why we couldn’t push the actual
Function instance onto the stack instead. I might try that once
everything else is working correctly, because an anonymous index on the
stack is going to be harder to debug & understand than an actual
Function instance.