Store table index (defaulting to zero) on element segment AST node
For now, it’s important to remember for a little while whether the table
index has been explicitly provided (i.e. by allowing it to be `nil` if
not), since that affects whether we’re parsing an active element segment
[0] and what abbreviations are available [1]. We eventually set it to
zero if it hasn’t been specified, since that’s the default.

[0] https://webassembly.github.io/spec/core/text/modules.html#element-segments
[1] https://webassembly.github.io/spec/core/text/modules.html#id7