Only try to parse inline element segment when know there is one
This conditional inside #parse_table tells us whether we’re using the
inline element segment abbrevation [0], and we only want to read `(elem
…)` in that case, so we should do that inside the conditional rather
than always doing it.

I’ve noticed that #parse_table_element is designed to silently tolerate
the absence of an `(elem …)`, which is no longer required, so I’ll fix
that in the next commit.

Annoyingly I couldn’t just use #reject in #initialise_tables to ignore
tables with no inline element segment because that would’ve thrown off
the indexes of the tables in the resulting array, so for now I’ve just
used `next` instead. Ideally I’ll find a way to deesugar the
abbreviations before they reach the interpreter, in which case all of
this complexity will go away anyway.

[0] https://webassembly.github.io/spec/core/text/modules.html#text-table-abbrev