Implement the `memory.copy` instruction
The specification uses a constant-space algorithm [0] which checks
whether the source address is before or after the destination address.
For the sake of simplicity I’ve ignored this and loaded a full copy of
the region before storing it; this might turn out to be too slow.

As with `memory.fill`, it’s probably desirable to implement Memory#copy
directly rather than doing it all through #load and #store, but I’m
focusing on making the tests pass right now.

[0] https://webassembly.github.io/spec/core/exec/instructions.html#xref-syntax-instructions-syntax-instr-memory-mathsf-memory-copy