Implement the `data.drop` instruction
The spec says this is “intended to be used as an optimization hint” and
“the memory used by this segment may be freed” [0], so we can in
principle free up the segment’s memory by replacing it with `nil` and
allowing it to be garbage collected. (I expect in practice that
everything is still hanging around as part of the AST, but that
hopefully won’t be true forever.)

To properly support dropped data segments we need to tolerate
`memory.init` when the length is zero, because that doesn’t actually try
to read the segment and therefore shouldn’t cause an exception even if
the segment has been dropped.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#memory-instructions