Implement the `elem.drop` instruction
The (virtually identical) implementation of `data.drop` required a
change to the implementation of `memory.init` to prevent it from trying
to access the dropped data segment in the case where the length was
zero. We don’t need to make the same change here to the implementation
of `table.init` because that implementation uses `length.times do … end`
which already does nothing in the zero-length case.