Add memory index space to identifier context
The spec says that the only valid memory index for a `data` segment is
zero [0], but that doesn’t necessarily mean that only a single memory
declaration is allowed, only that the first one must always be the
target of an active data segment. Still, I suspect there’s no practical
use for the memory index space. I’m adding it anyway since it makes the
rest of the implementation more regular, specifically the population of
the initial identifier context from explicit `import` fields.

[0] https://webassembly.github.io/spec/core/text/modules.html#data-segments