Add imported function/table/memory/global names to initial identifier context
This has finally been motivated by the desire to desugar inline imports
by the preprocessor so that the AST parser only needs to deal with
explicit `import` fields. That’s only possible if the identifiers from
the inline imports remain present in the initial context once desugared,
and for that to happen we need to correctly populate them during
the #build_initial_context method.