Deconstruct field incrementally in #process_field
In general, this “single pattern match” approach isn’t going to work for
other kinds of inline import (e.g. tables) because there may be a
variable number of list elements after the `(import …)` clause, and Ruby
pattern matching doesn’t give us an easy way to express that. By taking
the array apart in several pieces, we can be more specific about what
we’re expecting to see in each part of the abbreviation.