Expand inline table import abbreviations in preprocessor
It looks as though we never supported inline table imports anyway, but
the spec requires us to do so, and this new unit test checks that the
preprocessor is able to desugar them into (already-supported) explicit
`import` fields.

It’s fairly obvious that we’re going to want to combine the `func` and
`table` clauses in #process_field, but I’ll do that separately.