Include explicitly imported globals during module instantiation
This hasn’t happened before because we didn’t have a failing test for
it, but now that we’re trying to desugar inline global imports into
explicit `(import … (global …))`, the acceptance tests will fail unless
we have proper support support for imported globals.

The implementation of #initialise_globals needed to be hacked to
incorporate the count of imported globals when computing the index of
which global to initialise. This problem would go away if we were just
assigning initial values at the same time as we created the globals,
which we currently aren’t doing for the sake of letting Module be a Data
object rather than a Struct. There’s no real reason for it to be
immutable (provided we do in fact have so use for it to be mutable) so
we should probably just change this and make the code simpler.