Use two explicit pattern matches to deal with optional ID in #process_field
Although this introduces some repetition, I think it’s clearer and
avoids relying upon undefined behaviour of pattern matching (i.e. the
`rest` local does not get reassigned if the `in` match fails, which is
what MRI currently does but isn’t part of Ruby’s pattern matching
specification) and the subtle semantics of incrementally destructuring
an array with a combination of `=>` and `in`.

It would be really nice if Ruby’s pattern matching had better support
for this kind of thing, either by allowing the same variable to be bound
on both sides of an alternative pattern, or by explicitly supporting the
idea of “this element of an array pattern is optional”. But it doesn’t
yet, so for now I think the duplication is the least bad option, because
I’m prioritising clarity over concision and cleverness.