Expand inline module abbreviations in preprocessor
The top-level `(module …)` around a sequence of module fields can be
omitted entirely [0] so we need to add it back in as part of the
preprocessor. It’s clear from the test script grammar [1] that an inline
module could be followed by an assertion or similar, so I‘ve added a
unit test to make sure that situation is supported, even though
inline-module.wast doesn’t actually do it.

[0] https://webassembly.github.io/spec/core/text/modules.html#id10
[1] https://github.com/WebAssembly/spec/tree/main/interpreter#scripts