Support parsing `v128.const i64x2 0 0` in #parse_numeric_instruction
This is the only vector instruction which appears in test/core/*.wast
(and was only added yesterday [0]) so it’s not yet worth properly
implementing the instruction in general. The one occurrence in
linking.wast uses `i64x2` as its shape descriptor [1] so here we read
two numbers (actually zeros) after reading the instruction and shape.

[0] https://github.com/WebAssembly/spec/pull/1597
[1] https://webassembly.github.io/spec/core/syntax/instructions.html#syntax-vec-shape