Store a function’s parent module on its runtime representation
We need this so that we can properly set the current module when
evaluating the body of an imported function, all of whom’s identifiers
will refer to things inside the module where it was defined rather than
the module into which it’s been imported.

The hack in #initialise_functions is intended to set the current module
only if a function doesn’t already have one, since that would imply it’s
been imported from another module whose functions have already been
initialised. This is pretty grim and it’d be much better to just set
each function’s parent module when it is created, but at the moment we
don’t create a module until all of its functions have already been
built, so for the moment this is the easiest way to get the parent
module association working. In future we’ll need to go back and tidy
this up since it’s quite messy.

We initialise functions as early as possible in case module
instantiation fails during initialisation of globals, memory or tables.
This could potentially result in functions from this module being stored
in another module’s imported tables, so it’s important that those
functions are set up correctly first.