Desugar module definitions inside `assert_trap`
We need to be able to parse these so that we can attempt to instantiate
the module inside `assert_trap`, the side effects of which are tested by
linking.wast. But before we can ask the parser to parse `assert_trap`,
we need to make sure that any `(module …)` inside it has been desugared
ready for parsing.