Don’t try to initialise a table beyond its current length
This is important because elem.wast tries to use offsets like -1 which
turn into very large 32-bit numbers, and MRI completely freaks out if we
try to write to that index in the table elements array. It’s correct to
just raise an exception here because that’s exactly what the
`assert_trap` in elem.wast is expecting to happen.