Correctly convert integer arguments to unsigned integers
Integers are represented as unsigned numbers [0], so we need to do the
conversion before passing them to WebAssembly. There’s currently no
public API for doing this, so I’m having to call a private instance
method on Wasminna::Interpreter, which is obviously not ideal.

The purpose of all this test code is to help me to understand what
public API is required, so it’s okay to make a bit of a mess in the
short term as long as that mess gets refactored away as I create the API
I need.

[0] https://webassembly.github.io/spec/core/exec/numerics.html#integers