Descend into function definitions in preprocessor
We’re not expanding any abbreviations yet, but we need to know which
part of the function definition we’re inside before we’re able to deal
with things like typedefs and locals.

I’m not sure whether it’s best to destructively consume the function
definition here; we don’t yet have a consistent style throughout the
preprocessor. For now I’m going to focus on getting it working, and I’ll
make a decision about a coherent approach once all the pieces are there.