Descend into function bodies in preprocessor
An instruction can be a structured control instruction [0] which may
contain a type use (inside a block type [1]), or a plain `call_indirect`
instruction [2] which also may contain a type use. We need to reach both
of these places in order to expand any `param` or `result` abbreviation
that might exist there.

Structured control instructions may be nested to arbitrary depth, so we
descend recursively into instruction sequences until we hit individual
atoms. #process_instructions uses #flat_map in anticipation of the need
to expand an individual `(param …)` or `(result …)` into several, but
we’re not making use of that yet.

Function bodies aren’t the only place where instructions can appear, but
they’re the most obvious one, so we’ll start here and figure out the
rest later.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blockinstr
[1] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype
[2] https://webassembly.github.io/spec/core/text/instructions.html#text-plaininstr