Expect desugared syntax when parsing block type
At first glance this appears premature: the preprocessor so far only
desugars instructions which appear inside function bodies, but
instructions can appear [0] in global definitions, element segments
(including inline in a table definition) and data segments too.

However, validation requires instructions to be constant [1] in all
those places, and no form of constant instruction is able to
syntactically contain a block type, so no valid WebAssembly program will
ever need the `(param …)` or `(result …)` abbreviations to be desugared
there.

That doesn’t get us off the hook for preprocessing constant instructions
in general, it just means we don’t have to do it for this abbreviation
specifically. And that’s good because e.g. element segments themselves
support various other abbreviations [2] which we should probably tackle
first, and that’s a big distraction from the abbreviation we’re already
trying to implement.

So, for the sake of maintaining momentum, we’ll say that only supporting
valid programs (versus the larger set of syntactically correct but
invalid programs) is good enough right now.

[0] https://webassembly.github.io/spec/core/syntax/instructions.html#expressions
[1] https://webassembly.github.io/spec/core/valid/instructions.html#valid-constant
[2] https://webassembly.github.io/spec/core/text/modules.html#id7