Extract anonymous declarations helper in preprocessor
The preprocessor methods for `param`, `result` and `local` are virtually
identical because they all support the same multiple anonymous
occurrences abbreviation, so we can easily combine their
implementations.

The only behaviour change is that we’ll now silently tolerate a
(syntactically invalid) ID in a `result` instead of raising an
exception. That doesn’t matter for the preprocessor because its job is
just to expand abbreviations; the parser is responsible for checking
that the resulting syntax is actually valid.