Repurpose #read_lists as a declarations helper in AST parser
The #read_lists helper was only called by the three declaration-parsing
methods, so there’s not currently any need for it to be so general. We
can reduce repetition by specialising it for this single purpose.

We now fail slightly later if we find a `result` with an ID, but as long
as we still fail eventually, that should be fine.