Add current state and S-expression–reading helpers to preprocessor
This’ll allow us to incrementally consume the current S-expression (in
the same style as the AST parser) instead of passing it around between
methods as an argument. Mutation isn’t as conceptually tidy as a pure
functional solution but in practice it makes the whole implementation
more concise and easier to understand.