Lift #read_list helper out of #process_instructions
This turns #process_instructions into a method which operates entirely
upon the current S-expression state rather than taking an S-expression
as an argument.

The #process_instructions method was a leaf of the preprocessor’s call
graph, so now that we’ve converted it into operating on the current
state, we can convert the rest of the class’s implementation by
gradually working our way back up towards the root (#process_script).