Use optional ID in the result of #process_function
Instead of having two different (but very similiar) control flow paths
for functions with and without an ID, we can try to read an ID and then
unconditionally use the result when we build the final expression. If
there’s no ID then the result of reading it will be `nil`, but that
`nil` will just disappear when we splat it, so we get the same behaviour
as before.