Use S-expression helpers in #process_instructions
We can do this conversion because #process_instructions doesn’t call any
other methods which take an S-expression argument. (Well, except itself,
but we can paper over that crack temporarily by calling #s_expression
directly.) In the next commit we’ll lift the use of #read_list out to
the caller, leaving #process_instructions as a method which purely
operates on the current S-expression.