Combine basic field type handlers in #process_field
This is the most straightforward way of processing a field: look at what
type of field it is and call the corresponding method. We only need to
peek once at the first atom to make that decision.

It’s not ideal that we must first check the entire field to detect
inline imports and exports instead of consuming it incrementally with
`peek` and `read`, but that’s how this method works at the moment. I
intend to come back later and fix it; that’ll involve distributing the
handling of `import` and `export` abbreviations across the methods which
process the different types of field (`func`, `table`, `memory` and
`global`) where they can appear, which is very different from the change
I’m trying to make right now.

We’ll be able to stop wrapping every result in a one-element array once
we separate out the inline export handling, since that’s the only
operation which returns multiple fields.