Handle unchanged fields explicitly in #process_field
Previously, the do-nothing `else` clause applied in two different cases:

1. we’re processing a list (because fields are lists), but the list
   contains a field of a type that we don’t need to do any processing
   on; or

2. we’re processing a `binary` atom or one of the strings which follow
   it, and that’s not even a field.

The latter case is rare, should be treated separately, and ultimately
doesn’t even belong in #process_field. It’s not yet easy to move it out,
but we can at least separate it from the common former case to make the
distinction explicit.

The combination of `read_list` and `repeatedly { read }` is admittedly a
contrived way of writing the identity function, but it makes it obvious
we’re processing a list and puts the “unchanged field” clause into the
same shape as the other field clauses which will allow us to combine
them in future commits.