Read the next S-expression inside #process_field
Every caller now passes the value of `read` as an argument, so we can
safely push that responsibility down into the callee.

We can’t lift the #read_list call around the entire `case` because we
need to #peek at the full expression first to decide whether it contains
inline imports or exports.