Desugar missing table use abbreviation inside active element segment
We have to temporarily ignore the missing table use if the `func`
keyword is also missing, since some of the tests (e.g. bulk.wast) omit
both, and both abbreviations must be expanded together to maintain valid
syntax in that case. We’ll fix this when we implement the missing `func`
abbreviation.