Detect element list item type inside inline element segment
The expanded syntax for inline element segments [0] duplicates the
reftype in the resulting element list [1], but this should be replaced
by the `func` keyword when the element list contains function indexes so
that the function indexes abbreviation can be expanded correctly. We
distinguish that case by checking whether the first item is a list (i.e.
an element expression).

[0] https://webassembly.github.io/spec/core/text/modules.html#text-table-abbrev
[1] https://github.com/WebAssembly/spec/issues/1656