Process element segment separately depending upon mode
Active mode [0] is the hardest case to syntactically distinguish because
of all of the abbreviations it supports, but we do know that its ID must
be followed by either a table use or an offset, both of which are lists,
whereas declarative and passive mode must have either `declare`, a
reftype or (rarely) a function index, none of which are lists. So we
first look for a list to detect active mode, and otherwise use the
presence of `declare` to distinguish between declarative and passive
mode.

[0] https://webassembly.github.io/spec/core/syntax/modules.html#syntax-elem