Read active element segment reftype separately
This leaves only the unprocessed element list items remaining in the
element segment’s S-expression.

Note that `func` may be omitted in an active element segment only if the
table use is also omitted (“for backwards compatibility with earlier
versions of WebAssembly” [0]), so we have to be careful about the exact
circumstances under which we allow no reftype to be provided.

[0] https://webassembly.github.io/spec/core/text/modules.html#id7