Process instructions inside element expression
As with `offset`, the test is demonstrating desugaring of instructions
in an invalid program because that’s the correct behaviour for the
parsing phase regardless of validation. In practice only constant
instructions [0] may appear inside an element expression [1] but we have
no way to demonstrate desugaring of constant instructions yet.

[0] https://webassembly.github.io/spec/core/valid/instructions.html#valid-constant
[1] https://webassembly.github.io/spec/core/valid/modules.html#xref-syntax-modules-syntax-elem-mathsf-type-t-xref-syntax-modules-syntax-elem-mathsf-init-e-ast-xref-syntax-modules-syntax-elem-mathsf-mode-xref-syntax-modules-syntax-elemmode-mathit-elemmode