Use WebAssembly source code in preprocessor tests
The WebAssembly snippets are much easier to read and write than the Ruby
S-expression literals. That maintainability benefit makes up for the
fact that these unit tests are no longer fully isolated from the rest of
the system since they now depend upon the S-expression parser.

I experimented with storing JSON snapshots of known-good S-expression
parse results and relying on the snapshots in these tests instead of the
parser. In practice that felt like too much ceremony for too little
benefit. Parsing S-expressions is relatively simple so the small benefit
of isolating other unit tests from the parser is not worth the hassle of
maintaining the snapshots.