Return element list separately depending upon reftype
The implicit & explicit `func` cases support different abbreviations
from the `funcref` & `externref` case so they need to have independent
implementations. We can’t yet combine the implicit & explicit `func`
cases because of the need to carefully stage the desugaring of the
function indexes and missing `func` abbreviations [0] to avoid breaking
any of the existing tests.

[0] https://webassembly.github.io/spec/core/text/modules.html#id7