Desugar function indexes abbreviation inside element list
We can’t yet do this when `func` is entirely missing because that also
requires desugaring the missing `func` abbreviation.

Note that #process_function_indexes is, for the sake of clarity and
modularity, only expanding the abbreviation exactly as described in the
specification [0]. This requires the resulting element expressions to be
recursively desugared to add the missing `item`s.

[0] https://webassembly.github.io/spec/core/text/modules.html#id7