Process offset inside active data segment
We already implemented #process_offset as part of processing element
segments, so we can reuse it here, which automatically gives us support
for the missing `offset` abbreviation as well as (in principle)
desugaring the instructions inside the offset expression.

This leaves only the unprocessed strings remaining in the data segment’s
S-expression.