Lift #read_list helper to the top of #process_field
This does mean we have to temporarily match against the entire
S-expression instead of peeking at it, but that’ll go away once we start
consuming it incrementally. For now the priority is to get #read_list
lifted out of #process_field so that it’s operating entirely inside a
list like all the other methods do.