Preprocess instructions inside global definitions
As with `offset` and element expressions, the test is demonstrating
desugaring of instructions in an invalid program because that’s the
correct behaviour for the parsing phase regardless of validation. In
practice only constant instructions [0] may appear inside a global
definition [1] but we have no way to demonstrate desugaring of constant
instructions yet.

[0] https://webassembly.github.io/spec/core/valid/instructions.html#valid-constant
[1] https://webassembly.github.io/spec/core/valid/modules.html#globals