Organise tests with `BEGIN`/`END` blocks
execute_function_test.rb was already using `BEGIN` but float_test.rb was
organised in a different way (`def main`) and the other tests weren’t
really organised at all. There’s no good reason for them to be different
so this commit organises them all in the same style for consistency,
including using `END` to print the final newline.

The point of this, incidentally, is for each test file to be able to get
straight into the good stuff (i.e. assertions) without any setup getting
in the way.