Remove redundant `unless actual == expected` from tests
The condition is always false because it’s already inside the `else`
clause of an `if actual == expected` expression.

This was a silly oversight in 90ae714c4fed080f74a540ed4f2cd5819aa4dbd8
which I appear to have duplicated in all subsequent tests. Time to get
rid of it.