Extract Memory class from interpreter
The preprocessor will need to refer to `BYTES_PER_PAGE` when deciding
how many memory pages are needed to store an inline data segment. This
change allows the preprocessor to do that without depending upon the
interpreter.

There’s already a Wasminna::AST::Memory class so now we need to refer to
it more explicitly in the AST parser.