Introduce AST nodes for element segment modes
Every element segment has a mode — passive, active or declarative [0] —
and every segment of a particular mode contains a set of mode-specific
data. By introducing explicit AST nodes for each mode we’ll be able to
remove mode-specific data from the segment node, so data for a specific
mode will only be available on segments with that mode.

[0] https://webassembly.github.io/spec/core/syntax/modules.html#syntax-elem