Introduce AST nodes for data segment modes
Every data segment has a mode — passive or active [0] — and every
segment of a particular mode contains a set of mode-specific data. By
introducing explicit AST nodes for each mode we’ll be able to remove
mode-specific data from the segment node, so data for a specific mode
will only be available on segments with that mode.

Note that zero is the only valid memory index [1] in the current version
of WebAssembly, but we’re storing it explicitly in the active mode AST
node for the sake of clarity.

[0] https://webassembly.github.io/spec/core/syntax/modules.html#syntax-data
[1] https://webassembly.github.io/spec/core/valid/modules.html#valid-module