Extract ReadOptionalId helper module from preprocessor and parser
This logic is repeated in a lot of places, so let’s extract and
reuse it.

IDs are essentially always optional so I’ve chosen to write a single
method which unconditionally tries to read an ID (and may return nil)
rather than a pair of #can_read_id?/#read_id methods. That does mean we
have to check for `id.nil?` in a couple of places rather than asking a
dedicated predicate, but it’s rare enough that I think this trade-off is
the right one.