Insert a type index when a matching type definition already exists
This is the simplest case of the type use abbreviation [0]: if we can
find an existing type definition whose type is exactly the same as the
one being used, we insert a reference to that definition.

We’re not yet accounting for (irrelevant) parameter identifier
mismatches, nor for the case where no matching definition is found.

This commit includes a lot of tests. Earlier commits have already
plumbed the type definitions into every place where a type use can
syntactically occur, but this is our first opportunity to check that
it’s actually worked in all those cases.

[0] https://webassembly.github.io/spec/core/text/modules.html#abbreviations