Return deferred result from #process_passive_data_segment
This method is not a direct or indirect caller of #process_typeuse so
strictly speaking we don’t need it to return a deferred result, but it
makes the implementation of #process_data_segment tidier and more
regular if it does. We’ll err on the side of regularity and convert it
anyway for the sake of keeping #process_data_segment tidy.