Don’t treat `select` result type as a type use
The `select` instruction contains `result` types [0] which look
syntactically like type uses but explicitly aren’t in this context, so
we have to be careful to avoid triggering the type use preprocessor.

The tests in this commit enforce the constraint that a `select`
instruction’s `result` types must be left alone by any desugaring of the
type use abbreviation since they’re not type uses. We don’t have any
such desugaring yet, so the tests don’t currently constrain anything,
but they’ll prevent regressions when we do implement it.

[0] https://webassembly.github.io/spec/core/text/instructions.html#parametric-instructions