Return deferred result from #process_fields
This means the deferred result from #process_typeuse has reached
the #process_module method, which is where we expect it to be able to
get the full list of type definitions it needs once all of the module’s
fields have been processed.

Note that #process_fields converts an array of deferred results into a
deferred array of results by waiting until the list of type definitions
is available and then providing it to every element of the array.
Although it’s tempting to try to provide the type definitions to the
deferred results at this point (“after all fields”) rather than
returning a deferred result, we need #process_fields to be usable in
recursive calls within a module so we can’t assume we’re necessarily
processing _all_ of a module’s fields here.