Use #transpose to produce the array of fields in #process_fields
This’ll make it easier to also produce the array of type definitions in
a subsequent commit.

The call to #flat_map won’t ordinarily work if there are no fields to
process (`fields` would be nil because `[].transpose` is `[]` rather
than `[[], []]`) so we have to provide an empty array as a default.