Generate a new type definition if no matching one is found
Per the spec [0], “if no such index exists, then a new type definition
[…] is inserted at the end of the module”. We’re doing this by mutating
the incoming array of type definitions, which isn’t particularly elegant
but gets the job done without having to worry about threading another
result value (i.e. a separate array of generated type definitions)
through the field-processing methods.

We can’t add tests for this commit’s behaviour in isolation because the
resulting program is syntactically invalid: we’re inserting the index of
a type definition which hasn’t yet been added to the module. In the next
commit we’ll also insert the generated definitions at the end of the
module and add some tests to check that both halves of the solution are
working together correctly.

[0] https://webassembly.github.io/spec/core/text/modules.html#abbreviations