Return type definitions from #process_type_definition
To desugar the type use abbreviation [0] we’re going to need to know the
complete list of type definitions for the current module, so we must
collect them during processing.

The #process_type_definition method is where we expect the individual
type definitions to actually get generated. I’m using dummy ones for now
so we can focus on getting all of the callers plumbed together
correctly; we can decide the nature of the real values once we’re ready
to implement the desugaring.

This method now returns a pair of a deferred S-expression and a list of
type definitions. We’re going to need to push this pair up the call
graph in the same way as we did for deferred results so that the type
definitions reach #process_module where they’re needed.

As in bb63b2692fe88bd77dd209e1760049790d6f9309, we’ll begin by looking
at the static call graph of all methods which directly or indirectly
call #process_type_definition, starting from #process_module:

  #process_module
  └─ #process_fields
     └─ #process_field
        ├─ #process_function_definition
        │  └─ #expand_inline_import_export
        │     ├─ #expand_inline_import
        │     │  └─ (#process_fields…)
        │     └─ #expand_inline_export
        │        └─ (#process_fields…)
        ├─ #process_table_definition
        │  ├─ (#expand_inline_import_export…)
        │  └─ #expand_inline_element_segment
        │     └─ (#process_fields…)
        ├─ #process_memory_definition
        │  ├─ (#expand_inline_import_export…)
        │  └─ #expand_inline_data_segment
        │     └─ (#process_fields…)
        ├─ #process_global_definition
        │  └─ (#expand_inline_import_export…)
        └─ #process_type_definition

This allows us to organise the methods into distinct groups whose
members all share the same depth in the call graph (ignoring recursive
calls to #process_fields):

  * #expand_inline_import and #expand_inline_export (deepest)
  * #expand_inline_import_export, #expand_inline_element_segment and
    #expand_inline_data_segment
  * #process_function_definition, #process_table_definition,
    #process_memory_definition, #process_global_definition and
    #process_type_definition
  * #process_field
  * #process_fields
  * #process_module (shallowest)

By starting with the deepest methods and working our way up to the
shallowest, we’ll ensure that all of a method’s callees will have been
converted to return type definitions by the time we try to convert the
method itself, which allows us to make progress continuously upwards in
the call graph without needing to backtrack.

In the following commits we’ll work through the methods in this order
until the type definitions have reached #process_module.

[0] https://webassembly.github.io/spec/core/text/modules.html#abbreviations