Return type definitions from #process_import
This method is not a direct or indirect caller of #process_type_definitions
so strictly speaking we don’t need it to return any type definitions,
but it makes the implementation of #process_field tidier and more
regular if it does. We’ll err on the side of regularity and convert it
anyway for the sake of keeping #process_field tidy.