Update preprocessor tests to not depend upon type use abbreviation
Most of the preprocessor tests use syntax which relies upon the ability
to omit an explicit type use [0] and have it automatically inserted when
abbreviations are expanded. To prevent these tests from breaking when we
implement desugaring of type use abbreviations we need to update them to
contain full type uses which are unaffected by the expansion of that
abbreviation.

When adding an unabbreviated type use, I’ve removed the corresponding
inline declarations unless they’re the focus of the test.

[0] https://webassembly.github.io/spec/core/text/modules.html#abbreviations