Add generated type definitions to module fields in #process_module
We need a deep copy of the generated type definitions because they
contain object references to parameters and results elsewhere in the
source S-expression; since the AST parser consumes the S-expression
destructively, any structural sharing will cause serious problems.

The approach of “count the type definitions before and after processing
the fields, and the difference between these two counts must be the
newly-generated definitions” is a bit hacky, but it gives us the
behaviour we need. In future we should consider a more elegant approach
to this whole problem of modifying different parts of the S-expression
during preprocessing.