Preprocess pieces of block type before treating it as a type use
A block type isn’t a type use in certain circumstances [0] so we need to
read all of its parts first to make it possible to decide (in subsequent
commits) whether it’s actually a type use.

It’s important that we expand abbreviations before trying to decide
whether this is a type use or not. For example, `(param)` is an
abbreviation for no parameters, so the syntactic presence of `param`
doesn’t necessarily indicate a type use.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype