Add explicit `context:` argument to AST parser methods
The initial context is no longer mutated during parsing because the
preprocessor has already done the work to insert whatever type
definitions are needed, so we don’t need to worry about allowing methods
to update it. We’re therefore switching to an explicit argument since
it’s clearer and easier to understand.

For now we have to be careful to use the modified context (i.e.
`self.context`) inside #with_context blocks instead of the `context`
local variable. In the next commit we’ll be able to undo this when we
remove #with_context calls entirely.